stm32cube: f3: update to Cube version v1.11.6
Update the STM32CubeF3 package from version v1.11.5 to v1.11.6.
(Package source: https://github.com/STMicroelectronics/STM32CubeF3)

Additional modifications:
- use verbatim copyright text for `stm32_hal_legacy.h`
- drop undocumented patch to make LL_SetFlashLatency() public in
  `stm32f3xx_ll_utils.h`; Cube provides the function officially now
  (originally introduced in 6e2d3f951472597818416cfbfff9a601bc2e60e9)

+++ stm32cube/stm32f3xx/drivers/src/stm32f3xx_hal_rtc_ex.c
@@ -1672,7 +1672,7 @@ HAL_StatusTypeDef HAL_RTCEx_EnableBypassShadow(RTC_HandleTypeDef *hrtc)
   __HAL_RTC_WRITEPROTECTION_DISABLE(hrtc);
 
   /* Set the BYPSHAD bit */
-  hrtc->Instance->CR |= (uint8_t)RTC_CR_BYPSHAD;
+  hrtc->Instance->CR |= (uint32_t)RTC_CR_BYPSHAD;
 
   /* Enable the write protection for RTC registers */
   __HAL_RTC_WRITEPROTECTION_ENABLE(hrtc);
@@ -1705,7 +1705,7 @@ HAL_StatusTypeDef HAL_RTCEx_DisableBypassShadow(RTC_HandleTypeDef *hrtc)
   __HAL_RTC_WRITEPROTECTION_DISABLE(hrtc);
 
   /* Reset the BYPSHAD bit */
-  hrtc->Instance->CR &= (uint8_t)~RTC_CR_BYPSHAD;
+  hrtc->Instance->CR &= (uint32_t)~RTC_CR_BYPSHAD;
 
   /* Enable the write protection for RTC registers */
   __HAL_RTC_WRITEPROTECTION_ENABLE(hrtc);
